+++ changelog.md
@@ -1,3 +1,15 @@
+0.11.8
+------
+* add LandsatImage class
+
+0.11.7
+------
+* pin version of dask====0.13.0
+
+0.11.6
+------
+* Add optional index parameter to vector aggregations
+
 0.11.5
 ------
 * Add in vector aggregation support (see examples/vectors_aggregation.py)